Make service provider routing more Railsy
**Why**:
* No reason to nest inside `user` since we know who the current user is
via Devise
* This makes the routes more consistent and clean, imho
