+++ app/views/service_providers/edit.html.slim
@@ -1,0 +1,9 @@
+h2 = t('headings.service_providers.edit')
+
+= simple_form_for(service_provider, html: { autocomplete: 'off', role: 'form' }) do |form|
+  == render 'form', { form: form }
+  .wrapper.actions
+    .action
+      = form.button :submit, 'Update'
+    .action
+      = link_to t('forms.buttons.cancel'), service_provider_path(service_provider), class: 'usa-button usa-button-outline use-button-secondary', role: :button
