+++ app/views/user_mailer/admin_new_service_provider.html.slim
@@ -6,7 +6,7 @@ html
   p A new Identity service provider service_provider requires approval:
 
   p
-    =link_to @service_provider.friendly_name, edit_users_service_provider_url(@service_provider)
+    = link_to @service_provider.friendly_name, edit_service_provider_url(@service_provider)
 
   p
     strong PLEASE DO NOT REPLY TO THIS MESSAGE
