+++ app/views/user_mailer/user_approved_service_provider.html.slim
@@ -6,7 +6,7 @@ html
   p An Identity service provider service_provider has been approved.
 
   p
-    =link_to @service_provider.friendly_name, users_service_provider_url(@service_provider)
+    = link_to @service_provider.friendly_name, service_provider_url(@service_provider)
 
   p
     strong PLEASE DO NOT REPLY TO THIS MESSAGE
