+++ app/views/user_mailer/user_new_service_provider.html.slim
@@ -6,7 +6,7 @@ html
   p You have created a new Identity service provider:
 
   p
-    =link_to @service_provider.friendly_name, edit_users_service_provider_url(@service_provider)
+    =link_to @service_provider.friendly_name, edit_service_provider_url(@service_provider)
 
   p A message requesting approval has been sent to the site administrators.
